Jihyun

+++ src/main/java/com/project/model/ReplyDO.java
@@ -49,7 +49,7 @@ public class ReplyDO {
 		return post_id;
 	}
 
-	public void setPost_id(int post_id) {
+	public void setPost_id(long post_id) {
 		this.post_id = post_id;
 	}
